# openapi.yaml
openapi: 3.0.3
info:
  title: ALEXZA AI API
  description: |
    ALEXZA AI provides a managed AI runtime for running actions, managing projects,
    and receiving webhooks. Use `x-api-key` for API key authentication on v1 endpoints.
  version: 1.0.0
  contact:
    name: ALEXZA AI
servers:
  - url: /
    description: API server (use full base URL in production)
...
paths:
...
  /api/analytics/overview:
    get:
      tags: [Analytics]
      summary: Analytics overview
      description: Returns aggregated analytics for the user.
      operationId: getAnalyticsOverview
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  totalCreditsUsed:
                    type: number
                  totalApiCalls:
                    type: number
                  totalTokensUsed:
                    type: number
                  totalActionsRun:
                    type: number
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - SessionAuth: []
  /api/analytics/projects:
    get:
      tags: [Analytics]
      summary: Analytics by project
      operationId: getAnalyticsProjects
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  projects:
                    type: array
                    items:
                      type: object
                      properties:
                        projectId:
                          type: string
                        projectName:
                          type: string
                        totalApiCalls:
                          type: number
                        totalTokensUsed:
                          type: number
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - SessionAuth: []
  /api/analytics/actions:
    get:
      tags: [Analytics]
      summary: Analytics by action
      operationId: getAnalyticsActions
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  actions:
                    type: array
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - SessionAuth: []
  /api/analytics/daily:
    get:
      tags: [Analytics]
      summary: Daily analytics
      operationId: getAnalyticsDaily
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  daily:
                    type: array
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - SessionAuth: []
  /api/requests:
    get:
      tags: [Requests]
      summary: List API request logs
      description: Paginated list of API request history for debugging.
      operationId: listRequests
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: projectId
          in: query
          schema:
            type: string
        - name: actionName
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [success, error, failed_insufficient_credits]
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  requests:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiRequestItem"
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      pageSize:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - SessionAuth: []
  /api/requests/{id}:
    get:
      tags: [Requests]
      summary: Get request detail
      operationId: getRequestDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  request:
                    $ref: "#/components/schemas/ApiRequestItem"
        "401":
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
        - SessionAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key (e.g. axza_live_...)
    SessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie (set via login)
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          oneOf:
            - type: string
            - type: object
              properties:
                code:
                  type: string
                message:
                  type: string
        requestId:
          type: string
          description: Present on v1 run endpoints
...
    ApiRequestItem:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        projectName:
          type: string
        actionName:
          type: string
        status:
          type: string
          enum: [success, error, failed_insufficient_credits]
        tokensUsed:
          type: number
        latencyMs:
          type: number
        error:
          type: string
        createdAt:
          type: string
          format: date-time
...